+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Implement the `--version` argument in extractor.py.
 ### Changed
 * Further code cleanup.
+### Fixed
+* Specify a text encoding in `WebArchiveTest.test_webarchive_to_html()` so the test will pass on Windows.
 
 ## [0.3.1] - 2021-09-25
 ### Added